Configuración del sistema

+++ Workspace/Acme-Rendezvous/Acme-Rendezvous/src/main/java/services/AnnouncementService.java
@@ -4,7 +4,10 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Date;
 
+import javax.transaction.Transactional;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 import org.springframework.util.Assert;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.Validator;
@@ -12,8 +15,11 @@ import org.springframework.validation.Validator;
 import repositories.AnnouncementRepository;
 import domain.Announcement;
 import domain.Rendezvous;
+import domain.User;
 import forms.AnnouncementForm;
 
+@Service
+@Transactional
 public class AnnouncementService {
 
 	// Managed repository
@@ -24,6 +30,12 @@ public class AnnouncementService {
 	// Supporting services
 //	@Autowired
 //	private RendezvousService rendezvousService;
+	
+	@Autowired
+	private UserService userService;
+	
+	@Autowired
+	private AdminService adminService;
 
 	@Autowired
 	private Validator validator;
@@ -40,9 +52,9 @@ public class AnnouncementService {
 		Announcement res = new Announcement();
 		
 		Date moment = new Date(System.currentTimeMillis()-1);
-		Rendezvous rendezvous = new Rendezvous();
-		
-		res.setRendezvous(rendezvous);
+//		Rendezvous rendezvous = new Rendezvous();
+//		
+//		res.setRendezvous(rendezvous);
 		res.setMoment(moment);
 		
 		return res;
@@ -73,6 +85,23 @@ public class AnnouncementService {
 	public Announcement save(Announcement announcement){
 		Assert.notNull(announcement);
 		Announcement res;
+		User user = new User();
+		Collection<Rendezvous> rendezvousByPrincipal = new ArrayList<Rendezvous>();
+		Date moment;
+		
+		//editar sus propios announcements
+		if(announcement.getId() != 0){
+			Collection<Announcement> announcementUser;
+			announcementUser = this.announcementRepository.announcementsByUser(user.getId());
+			Assert.isTrue(announcementUser.contains(announcement));
+		}
+		
+		//crear announcements para los rendezvous que ha organizado
+		rendezvousByPrincipal = user.getOrganisedRendezvouses();
+		Assert.isTrue(rendezvousByPrincipal.contains(announcement.getRendezvous()));
+		
+		moment = new Date(System.currentTimeMillis()-1);
+		announcement.setMoment(moment);
 		
 		res = this.announcementRepository.save(announcement);
 		
@@ -81,19 +110,25 @@ public class AnnouncementService {
 	}
 	
 	public void delete(Announcement announcement){
+		this.adminService.checkAuthority();
+		
 		Assert.notNull(announcement);
 		Assert.isTrue(announcement.getId()!=0);
 		Assert.isTrue(this.announcementRepository.exists(announcement.getId()));
+//		
+//		Rendezvous rendezvous = new Rendezvous();
+//		rendezvous = announcement.getRendezvous();
+//		rendezvous.getAnnouncements().remove(announcement);
 		
 		this.announcementRepository.delete(announcement);
 	}
 	
 	// Other busines methods
 	
-	public Collection<Announcement> findAnnouncementByRendezvous(int idRendezvous){
+	public Collection<Announcement> findAnnouncementsByRendezvous(int rendezvousId){
 		Collection<Announcement> res = new ArrayList<Announcement>();
 		
-		res = announcementRepository.findAnnouncementByRendezvous(idRendezvous);
+		res = announcementRepository.findAnnouncementsByRendezvous(rendezvousId);
 		Assert.notNull(res);
 		
 		return res;
@@ -112,6 +147,15 @@ public class AnnouncementService {
 		
 		return res;
 	}
+	
+//	public Collection<Announcement> findAnnouncementsByAttendants(){
+//		Collection<Announcement> res;
+//		User user = new User();
+//		user = this.userService.findByPrincipal();
+//		res = this.announcementRepository.findAnnouncementsByAttendants(user.getId());
+//		Assert.notNull(res);
+//		return res;
+//	}
 
 
 }
